test(event): split event_test.go into per-domain files
The 3623-line event_test.go grew over time into a single alphabetized
blob covering every concern in the root package. Split into five
domain-coherent files alongside the slimmer event_test.go.

  event_test.go                2344 lines  (core event lifecycle,
                                            context, async, worker pool,
                                            transport, subscription,
                                            bus basics)
  bus_registry_test.go          304 lines  (TestGetBus, TestListBuses,
                                            TestDuplicateBusError, full-
                                            name resolution, etc.)
  middleware_test.go            613 lines  (bus-level idempotency /
                                            poison, schema-controlled
                                            middleware, schema timeout
                                            overrides, plus mock stores
                                            scoped to these tests)
  outbox_integration_test.go    153 lines  (TestOutboxIntegration,
                                            TestOutboxContextHelpers
                                            and the mockOutboxStore)
  message_filter_test.go        130 lines  (TestWithMessageFilter*,
                                            TestMessageFilter_*)
  decode_error_test.go          128 lines  (TestWithDecodeErrorHandler*,
                                            TestDecodeErrorHandler_*)

No behavior changes. Each file carries only the imports it needs.
The four mocks that the middleware/schema tests were the sole
consumers of (mockIdempotencyStore, mockPoisonDetector,
mockMonitorStore, mockSchemaProvider) move with those tests into
middleware_test.go; mockOutboxStore moves to outbox_integration_test.go.

go test -race -count=10 passes for the root package in ~12s (down
from ~17s).
