+++ event_test.go
@@ -2342,1282 +2342,3 @@ func TestTypedEventsSlice(t *testing.T) {
 	}
 }
 
-// =============================================================================
-// Bus Registry Tests
-// =============================================================================
-
-// TestGetBus verifies GetBus returns registered buses
-func TestGetBus(t *testing.T) {
-	t.Parallel()
-	busName := "test-getbus-" + NewID()
-	bus := mustNewBus(t, busName, WithTransport(channel.New()))
-	defer bus.Close(context.Background())
-
-	// GetBus should return the bus
-	got := GetBus(busName)
-	if got == nil {
-		t.Fatal("GetBus returned nil for registered bus")
-	}
-	if got.Name() != busName {
-		t.Errorf("expected bus name %q, got %q", busName, got.Name())
-	}
-
-	// GetBus for non-existent should return nil
-	if got := GetBus("non-existent-bus"); got != nil {
-		t.Errorf("expected nil for non-existent bus, got %v", got)
-	}
-}
-
-// TestListBuses verifies ListBuses returns all registered bus names
-func TestListBuses(t *testing.T) {
-	t.Parallel()
-	busName1 := "test-list-1-" + NewID()
-	busName2 := "test-list-2-" + NewID()
-
-	bus1 := mustNewBus(t, busName1, WithTransport(channel.New()))
-	defer bus1.Close(context.Background())
-
-	bus2 := mustNewBus(t, busName2, WithTransport(channel.New()))
-	defer bus2.Close(context.Background())
-
-	names := ListBuses()
-
-	found1, found2 := false, false
-	for _, name := range names {
-		if name == busName1 {
-			found1 = true
-		}
-		if name == busName2 {
-			found2 = true
-		}
-	}
-
-	if !found1 {
-		t.Errorf("ListBuses did not include %q", busName1)
-	}
-	if !found2 {
-		t.Errorf("ListBuses did not include %q", busName2)
-	}
-}
-
-// TestDuplicateBusError verifies NewBus returns error for duplicate name
-func TestDuplicateBusError(t *testing.T) {
-	t.Parallel()
-	busName := "test-duplicate-" + NewID()
-	bus1 := mustNewBus(t, busName, WithTransport(channel.New()))
-	defer bus1.Close(context.Background())
-
-	// Try to create another bus with same name
-	_, err := NewBus(busName, WithTransport(channel.New()))
-	if err == nil {
-		t.Fatal("expected error for duplicate bus name")
-	}
-	if !errors.Is(err, ErrBusExists) {
-		t.Errorf("expected ErrBusExists, got %v", err)
-	}
-}
-
-// TestBusUnregisteredOnClose verifies bus is removed from registry on Close
-func TestBusUnregisteredOnClose(t *testing.T) {
-	t.Parallel()
-	busName := "test-unregister-" + NewID()
-	bus := mustNewBus(t, busName, WithTransport(channel.New()))
-
-	// Bus should be registered
-	if GetBus(busName) == nil {
-		t.Fatal("bus not registered after creation")
-	}
-
-	// Close the bus
-	bus.Close(context.Background())
-
-	// Bus should no longer be registered
-	if GetBus(busName) != nil {
-		t.Fatal("bus still registered after Close")
-	}
-
-	// Should be able to create a new bus with same name
-	bus2, err := NewBus(busName, WithTransport(channel.New()))
-	if err != nil {
-		t.Fatalf("could not create bus after Close: %v", err)
-	}
-	defer bus2.Close(context.Background())
-}
-
-// TestGetEventByFullName verifies Get[T] works with full name
-func TestGetEventByFullName(t *testing.T) {
-	t.Parallel()
-	type Order struct {
-		ID string
-	}
-
-	busName := "test-fullname-" + NewID()
-	bus := mustNewBus(t, busName, WithTransport(channel.New()))
-	defer bus.Close(context.Background())
-
-	// Register event
-	ev := New[Order]("order.created")
-	if err := Register(context.Background(), bus, ev); err != nil {
-		t.Fatalf("failed to register event: %v", err)
-	}
-
-	// Get event by full name
-	fullName := busName + "://order.created"
-	event, err := Get[Order](fullName)
-	if err != nil {
-		t.Fatalf("Get[Order] failed: %v", err)
-	}
-	if event.Name() != "order.created" {
-		t.Errorf("expected event name 'order.created', got %q", event.Name())
-	}
-}
-
-// TestGetEventTypeMismatch verifies Get returns error for type mismatch
-func TestGetEventTypeMismatch(t *testing.T) {
-	t.Parallel()
-	type Order struct {
-		ID string
-	}
-	type User struct {
-		Name string
-	}
-
-	busName := "test-mismatch-" + NewID()
-	bus := mustNewBus(t, busName, WithTransport(channel.New()))
-	defer bus.Close(context.Background())
-
-	// Register event as Order
-	ev := New[Order]("order.created")
-	if err := Register(context.Background(), bus, ev); err != nil {
-		t.Fatalf("failed to register event: %v", err)
-	}
-
-	// Try to get as User - should fail
-	fullName := busName + "://order.created"
-	_, err := Get[User](fullName)
-	if err == nil {
-		t.Fatal("expected error for type mismatch")
-	}
-	if !errors.Is(err, ErrTypeMismatch) {
-		t.Errorf("expected ErrTypeMismatch, got %v", err)
-	}
-}
-
-// TestPublishByFullName verifies Publish works with full name
-func TestPublishByFullName(t *testing.T) {
-	t.Parallel()
-	type Order struct {
-		ID string
-	}
-
-	busName := "test-publish-fn-" + NewID()
-	bus := mustNewBus(t, busName, WithTransport(channel.New()))
-	defer bus.Close(context.Background())
-
-	// Register and subscribe
-	ev := New[Order]("order.created")
-	if err := Register(context.Background(), bus, ev); err != nil {
-		t.Fatalf("failed to register event: %v", err)
-	}
-
-	ch := make(chan Order)
-	ev.Subscribe(context.Background(), func(ctx context.Context, e Event[Order], order Order) error {
-		ch <- order
-		return nil
-	})
-
-	// Publish using full name
-	fullName := busName + "://order.created"
-	if err := Publish(context.Background(), fullName, Order{ID: "test-123"}); err != nil {
-		t.Fatalf("Publish failed: %v", err)
-	}
-
-	// Verify received
-	select {
-	case order := <-ch:
-		if order.ID != "test-123" {
-			t.Errorf("expected ID 'test-123', got %q", order.ID)
-		}
-	case <-time.After(100 * time.Millisecond):
-		t.Error("timeout waiting for event")
-	}
-}
-
-// TestSubscribeByFullName verifies Subscribe works with full name
-func TestSubscribeByFullName(t *testing.T) {
-	t.Parallel()
-	type Order struct {
-		ID string
-	}
-
-	busName := "test-subscribe-fn-" + NewID()
-	bus := mustNewBus(t, busName, WithTransport(channel.New()))
-	defer bus.Close(context.Background())
-
-	// Register event
-	ev := New[Order]("order.created")
-	if err := Register(context.Background(), bus, ev); err != nil {
-		t.Fatalf("failed to register event: %v", err)
-	}
-
-	// Subscribe using full name
-	ch := make(chan Order)
-	fullName := busName + "://order.created"
-	if err := Subscribe(context.Background(), fullName, func(ctx context.Context, e Event[Order], order Order) error {
-		ch <- order
-		return nil
-	}); err != nil {
-		t.Fatalf("Subscribe failed: %v", err)
-	}
-
-	// Publish and verify
-	ev.Publish(context.Background(), Order{ID: "test-456"})
-
-	select {
-	case order := <-ch:
-		if order.ID != "test-456" {
-			t.Errorf("expected ID 'test-456', got %q", order.ID)
-		}
-	case <-time.After(100 * time.Millisecond):
-		t.Error("timeout waiting for event")
-	}
-}
-
-// TestInvalidFullName verifies error handling for invalid full names
-func TestInvalidFullName(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		name     string
-		fullName string
-	}{
-		{"missing separator", "busname-eventname"},
-		{"empty bus name", "://eventname"},
-		{"empty event name", "busname://"},
-		{"no separator at all", "justsomestring"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := Get[any](tt.fullName)
-			if err == nil {
-				t.Errorf("expected error for full name %q", tt.fullName)
-			}
-			if !errors.Is(err, ErrInvalidFullName) && !errors.Is(err, ErrBusNotFound) {
-				t.Errorf("expected ErrInvalidFullName or ErrBusNotFound, got %v", err)
-			}
-		})
-	}
-}
-
-// TestGetEventNotFound verifies error for non-existent event
-func TestGetEventNotFound(t *testing.T) {
-	t.Parallel()
-	busName := "test-notfound-" + NewID()
-	bus := mustNewBus(t, busName, WithTransport(channel.New()))
-	defer bus.Close(context.Background())
-
-	// Try to get non-existent event
-	fullName := busName + "://nonexistent.event"
-	_, err := Get[any](fullName)
-	if err == nil {
-		t.Fatal("expected error for non-existent event")
-	}
-	if !errors.Is(err, ErrEventNotFound) {
-		t.Errorf("expected ErrEventNotFound, got %v", err)
-	}
-}
-
-// TestGetBusNotFound verifies error for non-existent bus in full name
-func TestGetBusNotFound(t *testing.T) {
-	t.Parallel()
-	fullName := "nonexistent-bus-12345://some.event"
-	_, err := Get[any](fullName)
-	if err == nil {
-		t.Fatal("expected error for non-existent bus")
-	}
-	if !errors.Is(err, ErrBusNotFound) {
-		t.Errorf("expected ErrBusNotFound, got %v", err)
-	}
-}
-
-// mockIdempotencyStore implements IdempotencyStore for testing
-type mockIdempotencyStore struct {
-	mu        sync.Mutex
-	processed map[string]bool
-}
-
-func newMockIdempotencyStore() *mockIdempotencyStore {
-	return &mockIdempotencyStore{processed: make(map[string]bool)}
-}
-
-func (s *mockIdempotencyStore) IsDuplicate(ctx context.Context, messageID string) (bool, error) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.processed[messageID], nil
-}
-
-func (s *mockIdempotencyStore) MarkProcessed(ctx context.Context, messageID string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.processed[messageID] = true
-	return nil
-}
-
-// mockPoisonDetector implements PoisonDetector for testing
-type mockPoisonDetector struct {
-	mu          sync.Mutex
-	failures    map[string]int
-	quarantined map[string]bool
-	threshold   int
-}
-
-func newMockPoisonDetector(threshold int) *mockPoisonDetector {
-	return &mockPoisonDetector{
-		failures:    make(map[string]int),
-		quarantined: make(map[string]bool),
-		threshold:   threshold,
-	}
-}
-
-func (d *mockPoisonDetector) Check(ctx context.Context, messageID string) (bool, error) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	return d.quarantined[messageID], nil
-}
-
-func (d *mockPoisonDetector) RecordFailure(ctx context.Context, messageID string) (bool, error) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	d.failures[messageID]++
-	if d.failures[messageID] >= d.threshold {
-		d.quarantined[messageID] = true
-		return true, nil
-	}
-	return false, nil
-}
-
-func (d *mockPoisonDetector) RecordSuccess(ctx context.Context, messageID string) error {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	delete(d.failures, messageID)
-	return nil
-}
-
-// TestBusLevelIdempotency verifies that bus-level idempotency automatically skips duplicates
-func TestBusLevelIdempotency(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-	idempStore := newMockIdempotencyStore()
-
-	bus := mustNewBus(t, "test-idemp-bus-"+randomString(5),
-		WithTransport(channel.New()),
-		WithIdempotency(idempStore),
-	)
-	defer bus.Close(ctx)
-
-	ev := New[string]("test.event")
-	if err := Register(ctx, bus, ev); err != nil {
-		t.Fatal(err)
-	}
-
-	var callCount atomic.Int32
-	ev.Subscribe(ctx, func(ctx context.Context, e Event[string], data string) error {
-		callCount.Add(1)
-		return nil
-	})
-
-	// Publish same message twice (same event ID via context)
-	msgCtx := ContextWithEventID(ctx, "msg-123")
-	ev.Publish(msgCtx, "hello")
-	eventuallyEqInt32(t, 2*time.Second, &callCount, 1, "first publish should be processed")
-
-	ev.Publish(msgCtx, "hello again")
-	consistentlyEqInt32(t, 100*time.Millisecond, &callCount, 1, "duplicate publish must not be processed")
-}
-
-// TestBusLevelPoisonDetection verifies that bus-level poison detection skips quarantined messages
-func TestBusLevelPoisonDetection(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-	poisonDetector := newMockPoisonDetector(2) // quarantine after 2 failures
-
-	bus := mustNewBus(t, "test-poison-bus-"+randomString(5),
-		WithTransport(channel.New()),
-		WithPoisonDetection(poisonDetector),
-	)
-	defer bus.Close(ctx)
-
-	ev := New[string]("test.poison.event")
-	if err := Register(ctx, bus, ev); err != nil {
-		t.Fatal(err)
-	}
-
-	var callCount atomic.Int32
-	ev.Subscribe(ctx, func(ctx context.Context, e Event[string], data string) error {
-		callCount.Add(1)
-		return errors.New("always fails")
-	})
-
-	// Publish message 3 times with same ID
-	msgCtx := ContextWithEventID(ctx, "poison-msg-456")
-	ev.Publish(msgCtx, "first")
-	eventuallyEqInt32(t, 2*time.Second, &callCount, 1, "first publish should be processed")
-
-	ev.Publish(msgCtx, "second")
-	eventuallyEqInt32(t, 2*time.Second, &callCount, 2, "second publish should trip the quarantine threshold")
-
-	ev.Publish(msgCtx, "third") // should be skipped (quarantined)
-	consistentlyEqInt32(t, 100*time.Millisecond, &callCount, 2, "quarantined message must not be processed")
-
-	// Verify message is quarantined
-	poisonDetector.mu.Lock()
-	isQuarantined := poisonDetector.quarantined["poison-msg-456"]
-	poisonDetector.mu.Unlock()
-	if !isQuarantined {
-		t.Error("expected message to be quarantined")
-	}
-}
-
-// TestBusLevelMiddlewareCombined verifies both idempotency and poison detection work together
-func TestBusLevelMiddlewareCombined(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-	idempStore := newMockIdempotencyStore()
-	poisonDetector := newMockPoisonDetector(3)
-
-	bus := mustNewBus(t, "test-combined-bus-"+randomString(5),
-		WithTransport(channel.New()),
-		WithIdempotency(idempStore),
-		WithPoisonDetection(poisonDetector),
-	)
-	defer bus.Close(ctx)
-
-	ev := New[string]("test.combined.event")
-	if err := Register(ctx, bus, ev); err != nil {
-		t.Fatal(err)
-	}
-
-	var callCount atomic.Int32
-	ev.Subscribe(ctx, func(ctx context.Context, e Event[string], data string) error {
-		callCount.Add(1)
-		return nil // success
-	})
-
-	// Publish two different messages
-	ev.Publish(ContextWithEventID(ctx, "msg-1"), "first")
-	eventuallyEqInt32(t, 2*time.Second, &callCount, 1, "first publish should be processed")
-
-	ev.Publish(ContextWithEventID(ctx, "msg-2"), "second")
-	eventuallyEqInt32(t, 2*time.Second, &callCount, 2, "second publish should be processed")
-
-	// Publish duplicate of first message
-	ev.Publish(ContextWithEventID(ctx, "msg-1"), "duplicate")
-	consistentlyEqInt32(t, 100*time.Millisecond, &callCount, 2, "duplicate publish must not be processed")
-}
-
-// mockMonitorStore implements MonitorStore for testing
-type mockMonitorStore struct {
-	mu      sync.Mutex
-	started map[string]bool
-	status  map[string]string
-}
-
-func newMockMonitorStore() *mockMonitorStore {
-	return &mockMonitorStore{
-		started: make(map[string]bool),
-		status:  make(map[string]string),
-	}
-}
-
-func (s *mockMonitorStore) RecordStart(ctx context.Context, params RecordStartParams) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.started[params.EventID] = true
-	return nil
-}
-
-func (s *mockMonitorStore) RecordComplete(ctx context.Context, params RecordCompleteParams) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.status[params.EventID] = params.Status
-	return nil
-}
-
-func (s *mockMonitorStore) wasRecorded(eventID string) bool {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.started[eventID]
-}
-
-// mockSchemaProvider implements SchemaProvider for testing
-type mockSchemaProvider struct {
-	mu      sync.RWMutex
-	schemas map[string]*EventSchema
-}
-
-func newMockSchemaProvider() *mockSchemaProvider {
-	return &mockSchemaProvider{
-		schemas: make(map[string]*EventSchema),
-	}
-}
-
-func (p *mockSchemaProvider) Get(ctx context.Context, eventName string) (*EventSchema, error) {
-	p.mu.RLock()
-	defer p.mu.RUnlock()
-	if schema, ok := p.schemas[eventName]; ok {
-		// Return a copy
-		copy := *schema
-		return &copy, nil
-	}
-	return nil, nil
-}
-
-func (p *mockSchemaProvider) Set(ctx context.Context, schema *EventSchema) error {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.schemas[schema.Name] = schema
-	return nil
-}
-
-func (p *mockSchemaProvider) Delete(ctx context.Context, eventName string) error {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	delete(p.schemas, eventName)
-	return nil
-}
-
-func (p *mockSchemaProvider) Watch(ctx context.Context) (<-chan SchemaChangeEvent, error) {
-	ch := make(chan SchemaChangeEvent, 100)
-	return ch, nil
-}
-
-func (p *mockSchemaProvider) List(ctx context.Context) ([]*EventSchema, error) {
-	p.mu.RLock()
-	defer p.mu.RUnlock()
-	result := make([]*EventSchema, 0, len(p.schemas))
-	for _, schema := range p.schemas {
-		copy := *schema
-		result = append(result, &copy)
-	}
-	return result, nil
-}
-
-func (p *mockSchemaProvider) Close() error {
-	return nil
-}
-
-// TestSchemaLoadingOnRegister verifies that schemas are loaded when events are registered
-func TestSchemaLoadingOnRegister(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-	provider := newMockSchemaProvider()
-
-	// Pre-register a schema
-	provider.Set(ctx, &EventSchema{
-		Name:              "order.created",
-		Version:           1,
-		SubTimeout:        5 * time.Second,
-		MaxRetries:        3,
-		EnableMonitor:     true,
-		EnableIdempotency: true,
-		EnablePoison:      false,
-	})
-
-	bus := mustNewBus(t, "test-schema-load-"+randomString(5),
-		WithTransport(channel.New()),
-		WithSchemaProvider(provider),
-	)
-	defer bus.Close(ctx)
-
-	ev := New[string]("order.created")
-	if err := Register(ctx, bus, ev); err != nil {
-		t.Fatal(err)
-	}
-
-	// Verify schema was loaded by checking internal state
-	impl := ev.(*eventImpl[string])
-	if !impl.schema.loaded {
-		t.Error("expected schema to be loaded")
-	}
-	if !impl.schema.enableMonitor {
-		t.Error("expected enableMonitor to be true")
-	}
-	if !impl.schema.enableIdempotency {
-		t.Error("expected enableIdempotency to be true")
-	}
-	if impl.schema.enablePoison {
-		t.Error("expected enablePoison to be false")
-	}
-	if impl.subTimeout != 5*time.Second {
-		t.Errorf("expected subTimeout 5s, got %v", impl.subTimeout)
-	}
-	if impl.maxRetries != 3 {
-		t.Errorf("expected maxRetries 3, got %d", impl.maxRetries)
-	}
-}
-
-// TestSchemaNotFoundFallback verifies that missing schema doesn't prevent registration
-func TestSchemaNotFoundFallback(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-	provider := newMockSchemaProvider()
-	// Don't pre-register any schema
-
-	bus := mustNewBus(t, "test-schema-notfound-"+randomString(5),
-		WithTransport(channel.New()),
-		WithSchemaProvider(provider),
-	)
-	defer bus.Close(ctx)
-
-	ev := New[string]("order.created")
-	if err := Register(ctx, bus, ev); err != nil {
-		t.Fatal(err)
-	}
-
-	// Verify schema was not loaded
-	impl := ev.(*eventImpl[string])
-	if impl.schema.loaded {
-		t.Error("expected schema NOT to be loaded")
-	}
-}
-
-// TestSchemaControlsMiddleware verifies that schema flags control which middleware is applied
-func TestSchemaControlsMiddleware(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-
-	t.Run("schema enables only monitor", func(t *testing.T) {
-		provider := newMockSchemaProvider()
-		provider.Set(ctx, &EventSchema{
-			Name:              "test.event",
-			Version:           1,
-			EnableMonitor:     true,
-			EnableIdempotency: false,
-			EnablePoison:      false,
-		})
-
-		idempStore := newMockIdempotencyStore()
-		poisonDetector := newMockPoisonDetector(2)
-		monitorStore := newMockMonitorStore()
-
-		bus := mustNewBus(t, "test-schema-monitor-"+randomString(5),
-			WithTransport(channel.New()),
-			WithSchemaProvider(provider),
-			WithIdempotency(idempStore),
-			WithPoisonDetection(poisonDetector),
-			WithMonitor(monitorStore),
-		)
-		defer bus.Close(ctx)
-
-		ev := New[string]("test.event")
-		if err := Register(ctx, bus, ev); err != nil {
-			t.Fatal(err)
-		}
-
-		var received atomic.Bool
-		ev.Subscribe(ctx, func(ctx context.Context, e Event[string], data string) error {
-			received.Store(true)
-			return nil
-		})
-
-		msgID := "test-msg-" + randomString(5)
-		ev.Publish(ContextWithEventID(ctx, msgID), "hello")
-		eventuallyTrue(t, 2*time.Second, received.Load, "handler should have been called")
-
-		// Monitor should have recorded (enabled in schema)
-		eventuallyTrue(t, 2*time.Second, func() bool { return monitorStore.wasRecorded(msgID) },
-			"monitor should have recorded the event")
-
-		// Idempotency store should NOT have been called (disabled in schema)
-		idempStore.mu.Lock()
-		_, wasProcessed := idempStore.processed[msgID]
-		idempStore.mu.Unlock()
-		if wasProcessed {
-			t.Error("idempotency should NOT have been applied (disabled in schema)")
-		}
-	})
-
-	t.Run("schema enables only idempotency", func(t *testing.T) {
-		provider := newMockSchemaProvider()
-		provider.Set(ctx, &EventSchema{
-			Name:              "test.idemp.event",
-			Version:           1,
-			EnableMonitor:     false,
-			EnableIdempotency: true,
-			EnablePoison:      false,
-		})
-
-		idempStore := newMockIdempotencyStore()
-		monitorStore := newMockMonitorStore()
-
-		bus := mustNewBus(t, "test-schema-idemp-"+randomString(5),
-			WithTransport(channel.New()),
-			WithSchemaProvider(provider),
-			WithIdempotency(idempStore),
-			WithMonitor(monitorStore),
-		)
-		defer bus.Close(ctx)
-
-		ev := New[string]("test.idemp.event")
-		if err := Register(ctx, bus, ev); err != nil {
-			t.Fatal(err)
-		}
-
-		var callCount atomic.Int32
-		ev.Subscribe(ctx, func(ctx context.Context, e Event[string], data string) error {
-			callCount.Add(1)
-			return nil
-		})
-
-		msgID := "test-msg-" + randomString(5)
-		// Publish same message twice
-		ev.Publish(ContextWithEventID(ctx, msgID), "hello")
-		eventuallyEqInt32(t, 2*time.Second, &callCount, 1, "first publish should be processed")
-		ev.Publish(ContextWithEventID(ctx, msgID), "hello again")
-		consistentlyEqInt32(t, 100*time.Millisecond, &callCount, 1, "duplicate must be skipped by idempotency")
-
-		// Monitor should NOT have recorded (disabled in schema)
-		if monitorStore.wasRecorded(msgID) {
-			t.Error("monitor should NOT have recorded (disabled in schema)")
-		}
-	})
-}
-
-// TestNoSchemaFallbackToBusMiddleware verifies middleware is applied when no schema exists
-func TestNoSchemaFallbackToBusMiddleware(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-	provider := newMockSchemaProvider()
-	// No schema registered
-
-	idempStore := newMockIdempotencyStore()
-	monitorStore := newMockMonitorStore()
-
-	bus := mustNewBus(t, "test-no-schema-fallback-"+randomString(5),
-		WithTransport(channel.New()),
-		WithSchemaProvider(provider),
-		WithIdempotency(idempStore),
-		WithMonitor(monitorStore),
-	)
-	defer bus.Close(ctx)
-
-	ev := New[string]("unregistered.event")
-	if err := Register(ctx, bus, ev); err != nil {
-		t.Fatal(err)
-	}
-
-	var callCount atomic.Int32
-	ev.Subscribe(ctx, func(ctx context.Context, e Event[string], data string) error {
-		callCount.Add(1)
-		return nil
-	})
-
-	msgID := "test-msg-" + randomString(5)
-	// Publish same message twice
-	ev.Publish(ContextWithEventID(ctx, msgID), "hello")
-	eventuallyEqInt32(t, 2*time.Second, &callCount, 1, "first publish should be processed")
-	ev.Publish(ContextWithEventID(ctx, msgID), "hello again")
-	consistentlyEqInt32(t, 100*time.Millisecond, &callCount, 1, "fallback idempotency must skip duplicate")
-
-	// Monitor should have recorded (fallback behavior)
-	eventuallyTrue(t, 2*time.Second, func() bool { return monitorStore.wasRecorded(msgID) },
-		"monitor should have recorded (fallback behavior)")
-}
-
-// TestSchemaDisablesAllMiddleware verifies that schema can disable all middleware
-func TestSchemaDisablesAllMiddleware(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-	provider := newMockSchemaProvider()
-	provider.Set(ctx, &EventSchema{
-		Name:              "test.no.middleware",
-		Version:           1,
-		EnableMonitor:     false,
-		EnableIdempotency: false,
-		EnablePoison:      false,
-	})
-
-	idempStore := newMockIdempotencyStore()
-	poisonDetector := newMockPoisonDetector(2)
-	monitorStore := newMockMonitorStore()
-
-	bus := mustNewBus(t, "test-schema-disable-all-"+randomString(5),
-		WithTransport(channel.New()),
-		WithSchemaProvider(provider),
-		WithIdempotency(idempStore),
-		WithPoisonDetection(poisonDetector),
-		WithMonitor(monitorStore),
-	)
-	defer bus.Close(ctx)
-
-	ev := New[string]("test.no.middleware")
-	if err := Register(ctx, bus, ev); err != nil {
-		t.Fatal(err)
-	}
-
-	var callCount atomic.Int32
-	ev.Subscribe(ctx, func(ctx context.Context, e Event[string], data string) error {
-		callCount.Add(1)
-		return nil
-	})
-
-	msgID := "test-msg-" + randomString(5)
-	// Publish same message twice
-	ev.Publish(ContextWithEventID(ctx, msgID), "hello")
-	eventuallyEqInt32(t, 2*time.Second, &callCount, 1, "first publish should be processed")
-	ev.Publish(ContextWithEventID(ctx, msgID), "hello again")
-	eventuallyEqInt32(t, 2*time.Second, &callCount, 2, "duplicate must be processed when all middleware disabled")
-
-	// Monitor should NOT have recorded
-	if monitorStore.wasRecorded(msgID) {
-		t.Error("monitor should NOT have recorded (disabled in schema)")
-	}
-
-	// Idempotency should NOT have been applied
-	idempStore.mu.Lock()
-	_, wasProcessed := idempStore.processed[msgID]
-	idempStore.mu.Unlock()
-	if wasProcessed {
-		t.Error("idempotency should NOT have been applied (disabled in schema)")
-	}
-}
-
-// TestSchemaTimeoutApplied verifies that schema timeout is applied to events
-func TestSchemaTimeoutApplied(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-	provider := newMockSchemaProvider()
-	provider.Set(ctx, &EventSchema{
-		Name:       "test.timeout",
-		Version:    1,
-		SubTimeout: 100 * time.Millisecond,
-	})
-
-	bus := mustNewBus(t, "test-schema-timeout-"+randomString(5),
-		WithTransport(channel.New()),
-		WithSchemaProvider(provider),
-	)
-	defer bus.Close(ctx)
-
-	// Event without timeout option - should get timeout from schema
-	ev := New[string]("test.timeout")
-	if err := Register(ctx, bus, ev); err != nil {
-		t.Fatal(err)
-	}
-
-	impl := ev.(*eventImpl[string])
-	if impl.subTimeout != 100*time.Millisecond {
-		t.Errorf("expected subTimeout 100ms from schema, got %v", impl.subTimeout)
-	}
-}
-
-// TestEventTimeoutOverridesSchema verifies that event-level timeout takes precedence
-func TestEventTimeoutOverridesSchema(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-	provider := newMockSchemaProvider()
-	provider.Set(ctx, &EventSchema{
-		Name:       "test.timeout.override",
-		Version:    1,
-		SubTimeout: 100 * time.Millisecond,
-	})
-
-	bus := mustNewBus(t, "test-schema-timeout-override-"+randomString(5),
-		WithTransport(channel.New()),
-		WithSchemaProvider(provider),
-	)
-	defer bus.Close(ctx)
-
-	// Event WITH timeout option - should keep event timeout
-	ev := New[string]("test.timeout.override", WithSubscriberTimeout(500*time.Millisecond))
-	if err := Register(ctx, bus, ev); err != nil {
-		t.Fatal(err)
-	}
-
-	impl := ev.(*eventImpl[string])
-	// Event timeout should be preserved (schema doesn't override existing values)
-	if impl.subTimeout != 500*time.Millisecond {
-		t.Errorf("expected subTimeout 500ms from event option, got %v", impl.subTimeout)
-	}
-}
-
-// mockOutboxStore is a mock implementation of OutboxStore for testing
-type mockOutboxStore struct {
-	mu       sync.Mutex
-	messages []mockOutboxMessage
-}
-
-type mockOutboxMessage struct {
-	eventName string
-	eventID   string
-	payload   []byte
-	metadata  map[string]string
-}
-
-func (m *mockOutboxStore) Store(ctx context.Context, eventName string, eventID string, payload []byte, metadata map[string]string) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.messages = append(m.messages, mockOutboxMessage{
-		eventName: eventName,
-		eventID:   eventID,
-		payload:   payload,
-		metadata:  metadata,
-	})
-	return nil
-}
-
-func (m *mockOutboxStore) Messages() []mockOutboxMessage {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	return append([]mockOutboxMessage{}, m.messages...)
-}
-
-func TestOutboxIntegration(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-
-	// Create mock outbox store
-	outbox := &mockOutboxStore{}
-
-	// Create bus with outbox
-	bus := mustNewBus(t, faker.Word(), WithTransport(channel.New()), WithOutbox(outbox))
-	defer bus.Close(context.Background())
-
-	// Create and register event
-	orderEvent := New[string]("order.created")
-	if err := Register(ctx, bus, orderEvent); err != nil {
-		t.Fatal(err)
-	}
-
-	t.Run("publish without transaction goes to transport", func(t *testing.T) {
-		// Track received via subscriber
-		received := make(chan string, 1)
-		err := orderEvent.Subscribe(ctx, func(ctx context.Context, ev Event[string], data string) error {
-			received <- data
-			return nil
-		})
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// Publish without transaction context
-		err = orderEvent.Publish(ctx, "order-123")
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// Should receive via transport
-		select {
-		case data := <-received:
-			if data != "order-123" {
-				t.Errorf("expected order-123, got %s", data)
-			}
-		case <-time.After(100 * time.Millisecond):
-			t.Error("expected to receive message via transport")
-		}
-
-		// Outbox should be empty
-		if len(outbox.Messages()) != 0 {
-			t.Errorf("expected no messages in outbox, got %d", len(outbox.Messages()))
-		}
-	})
-
-	t.Run("publish inside transaction goes to outbox", func(t *testing.T) {
-		// Create transaction context
-		txCtx := WithOutboxTx(ctx, "mock-session")
-
-		// Publish inside transaction
-		err := orderEvent.Publish(txCtx, "order-456")
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// Message should be in outbox
-		messages := outbox.Messages()
-		if len(messages) != 1 {
-			t.Fatalf("expected 1 message in outbox, got %d", len(messages))
-		}
-
-		msg := messages[0]
-		if msg.eventName != "order.created" {
-			t.Errorf("expected event name order.created, got %s", msg.eventName)
-		}
-		if !strings.Contains(string(msg.payload), "order-456") {
-			t.Errorf("expected payload to contain order-456, got %s", string(msg.payload))
-		}
-	})
-}
-
-func TestOutboxContextHelpers(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-
-	t.Run("InOutboxTx returns false without transaction", func(t *testing.T) {
-		if InOutboxTx(ctx) {
-			t.Error("expected InOutboxTx to return false")
-		}
-	})
-
-	t.Run("InOutboxTx returns true with transaction", func(t *testing.T) {
-		txCtx := WithOutboxTx(ctx, "mock-session")
-		if !InOutboxTx(txCtx) {
-			t.Error("expected InOutboxTx to return true")
-		}
-	})
-
-	t.Run("OutboxTx returns session", func(t *testing.T) {
-		session := "mock-session"
-		txCtx := WithOutboxTx(ctx, session)
-		got := OutboxTx(txCtx)
-		if got != session {
-			t.Errorf("expected session %v, got %v", session, got)
-		}
-	})
-
-	t.Run("OutboxTx returns nil without transaction", func(t *testing.T) {
-		got := OutboxTx(ctx)
-		if got != nil {
-			t.Errorf("expected nil, got %v", got)
-		}
-	})
-}
-
-func TestWithMessageFilter(t *testing.T) {
-	t.Parallel()
-	filter := func(meta map[string]string) bool {
-		return meta["collection"] == "orders"
-	}
-	opts := newEventOptions(WithMessageFilter(filter))
-	if opts.messageFilter == nil {
-		t.Fatal("expected messageFilter to be set")
-	}
-	// Verify the filter works correctly
-	if !opts.messageFilter(map[string]string{"collection": "orders"}) {
-		t.Error("expected filter to return true for matching metadata")
-	}
-	if opts.messageFilter(map[string]string{"collection": "users"}) {
-		t.Error("expected filter to return false for non-matching metadata")
-	}
-}
-
-func TestWithMessageFilter_Nil(t *testing.T) {
-	t.Parallel()
-	opts := newEventOptions(WithMessageFilter(nil))
-	if opts.messageFilter != nil {
-		t.Error("expected nil messageFilter to be stored as nil")
-	}
-}
-
-func TestMessageFilter_SkipsNonMatching(t *testing.T) {
-	t.Parallel()
-	bus := mustNewBus(t, "test-filter", WithTransport(channel.New()))
-	defer bus.Close(context.Background())
-
-	received := make(chan string, 10)
-
-	e := New[string]("test-filtered",
-		WithMessageFilter(func(meta map[string]string) bool {
-			return meta["collection"] == "orders"
-		}),
-	)
-	if err := Register(context.Background(), bus, e); err != nil {
-		t.Fatalf("failed to register event: %v", err)
-	}
-
-	e.Subscribe(context.Background(), func(ctx context.Context, ev Event[string], data string) error {
-		received <- data
-		return nil
-	})
-
-	// Publish with matching metadata — should be received
-	ctx1 := ContextWithMetadata(context.Background(), map[string]string{"collection": "orders"})
-	if err := e.Publish(ctx1, "order-1"); err != nil {
-		t.Fatalf("publish failed: %v", err)
-	}
-
-	// Publish with non-matching metadata — should be filtered out
-	ctx2 := ContextWithMetadata(context.Background(), map[string]string{"collection": "users"})
-	if err := e.Publish(ctx2, "user-1"); err != nil {
-		t.Fatalf("publish failed: %v", err)
-	}
-
-	// Publish another matching one to confirm ordering
-	ctx3 := ContextWithMetadata(context.Background(), map[string]string{"collection": "orders"})
-	if err := e.Publish(ctx3, "order-2"); err != nil {
-		t.Fatalf("publish failed: %v", err)
-	}
-
-	// Should receive "order-1"
-	data, ok := waitForData(received, waitChTimeoutMS)
-	if !ok {
-		t.Fatal("expected to receive order-1")
-	}
-	if data != "order-1" {
-		t.Errorf("expected order-1, got %s", data)
-	}
-
-	// Should receive "order-2"
-	data, ok = waitForData(received, waitChTimeoutMS)
-	if !ok {
-		t.Fatal("expected to receive order-2")
-	}
-	if data != "order-2" {
-		t.Errorf("expected order-2, got %s", data)
-	}
-
-	// Should NOT receive anything else
-	_, ok = waitForData(received, waitChTimeoutMS)
-	if ok {
-		t.Error("expected no more messages, but received one")
-	}
-}
-
-func TestMessageFilter_NilAcceptsAll(t *testing.T) {
-	t.Parallel()
-	bus := mustNewBus(t, "test-filter-nil", WithTransport(channel.New()))
-	defer bus.Close(context.Background())
-
-	received := make(chan string, 10)
-
-	// No message filter — all messages should pass through
-	e := New[string]("test-no-filter")
-	if err := Register(context.Background(), bus, e); err != nil {
-		t.Fatalf("failed to register event: %v", err)
-	}
-
-	e.Subscribe(context.Background(), func(ctx context.Context, ev Event[string], data string) error {
-		received <- data
-		return nil
-	})
-
-	ctx := ContextWithMetadata(context.Background(), map[string]string{"collection": "anything"})
-	if err := e.Publish(ctx, "msg-1"); err != nil {
-		t.Fatalf("publish failed: %v", err)
-	}
-
-	data, ok := waitForData(received, waitChTimeoutMS)
-	if !ok {
-		t.Fatal("expected to receive msg-1")
-	}
-	if data != "msg-1" {
-		t.Errorf("expected msg-1, got %s", data)
-	}
-}
-
-func TestWithDecodeErrorHandler(t *testing.T) {
-	t.Parallel()
-	handler := func(ctx context.Context, msg message.Message, err error) error {
-		return ErrReject
-	}
-	opts := newEventOptions(WithDecodeErrorHandler(handler))
-	if opts.decodeErrorHandler == nil {
-		t.Fatal("expected decodeErrorHandler to be set")
-	}
-}
-
-func TestWithDecodeErrorHandler_Nil(t *testing.T) {
-	t.Parallel()
-	opts := newEventOptions(WithDecodeErrorHandler(nil))
-	if opts.decodeErrorHandler != nil {
-		t.Error("expected nil decodeErrorHandler to be stored as nil")
-	}
-}
-
-func TestDecodeErrorHandler_AckSkipsDLQ(t *testing.T) {
-	t.Parallel()
-	dlqStore := newTestDLQStore()
-	bus := mustNewBus(t, "test-decode-ack", WithTransport(channel.New()), WithDLQ(dlqStore))
-	defer bus.Close(context.Background())
-
-	handlerCalled := make(chan struct{}, 1)
-
-	e := New[string]("test-decode-ack",
-		WithDecodeErrorHandler(func(ctx context.Context, msg message.Message, err error) error {
-			handlerCalled <- struct{}{}
-			return nil // Ack — skip DLQ
-		}),
-	)
-	if err := Register(context.Background(), bus, e); err != nil {
-		t.Fatalf("failed to register event: %v", err)
-	}
-
-	e.Subscribe(context.Background(), func(ctx context.Context, ev Event[string], data string) error {
-		t.Error("handler should not be called on decode error")
-		return nil
-	})
-
-	// Send malformed JSON payload directly via bus
-	metadata := map[string]string{MetadataContentType: "application/json"}
-	if err := bus.Send(context.Background(), "test-decode-ack", "", []byte("{invalid json}"), metadata); err != nil {
-		t.Fatalf("send failed: %v", err)
-	}
-
-	if !wait(handlerCalled, waitChTimeoutMS) {
-		t.Fatal("expected decode error handler to be called")
-	}
-
-	// DLQ must NOT be called when the decode error handler returns nil.
-	// Wait a window long enough to catch a stray async DLQ enqueue.
-	select {
-	case <-dlqStore.called:
-		t.Error("expected DLQ store NOT to be called when decode error handler returns nil")
-	case <-time.After(50 * time.Millisecond):
-	}
-}
-
-func TestDecodeErrorHandler_RejectSendsToDLQ(t *testing.T) {
-	t.Parallel()
-	dlqStore := newTestDLQStore()
-	bus := mustNewBus(t, "test-decode-reject", WithTransport(channel.New()), WithDLQ(dlqStore))
-	defer bus.Close(context.Background())
-
-	e := New[string]("test-decode-reject",
-		WithDecodeErrorHandler(func(ctx context.Context, msg message.Message, err error) error {
-			return ErrReject // Send to DLQ
-		}),
-	)
-	if err := Register(context.Background(), bus, e); err != nil {
-		t.Fatalf("failed to register event: %v", err)
-	}
-
-	e.Subscribe(context.Background(), func(ctx context.Context, ev Event[string], data string) error {
-		t.Error("handler should not be called on decode error")
-		return nil
-	})
-
-	metadata := map[string]string{MetadataContentType: "application/json"}
-	if err := bus.Send(context.Background(), "test-decode-reject", "", []byte("{invalid json}"), metadata); err != nil {
-		t.Fatalf("send failed: %v", err)
-	}
-
-	if !wait(dlqStore.called, waitChTimeoutMS) {
-		t.Fatal("expected DLQ store to be called when decode error handler returns ErrReject")
-	}
-}
-
-func TestDecodeErrorHandler_DefaultBehaviorWhenNotSet(t *testing.T) {
-	t.Parallel()
-	dlqStore := newTestDLQStore()
-	bus := mustNewBus(t, "test-decode-default", WithTransport(channel.New()), WithDLQ(dlqStore))
-	defer bus.Close(context.Background())
-
-	// No WithDecodeErrorHandler — should use default DLQ+ack behavior
-	e := New[string]("test-decode-default")
-	if err := Register(context.Background(), bus, e); err != nil {
-		t.Fatalf("failed to register event: %v", err)
-	}
-
-	e.Subscribe(context.Background(), func(ctx context.Context, ev Event[string], data string) error {
-		t.Error("handler should not be called on decode error")
-		return nil
-	})
-
-	metadata := map[string]string{MetadataContentType: "application/json"}
-	if err := bus.Send(context.Background(), "test-decode-default", "", []byte("{invalid json}"), metadata); err != nil {
-		t.Fatalf("send failed: %v", err)
-	}
-
-	if !wait(dlqStore.called, waitChTimeoutMS) {
-		t.Fatal("expected DLQ store to be called with default behavior (no decode error handler set)")
-	}
-}
